+++ src/logic/ICity.ts
@@ -1,6 +1,6 @@
 import { CityName } from "./CityName";
-import { IWare } from "./i-ware";
 import { IInventory } from "./IInventory";
+import { IWare } from "./IWare";
 import { WareType } from "./wareType";
 
 export interface ICity extends IInventory {
