+++ src/logic/IWare.ts
