+++ src/logic/IWareForCity.ts
@@ -1,4 +1,4 @@
-import { IWare } from "./i-ware";
+import { IWare } from "./IWare";
 
 export interface IWareForCity extends IWare {
     readonly maxPrice: number;
